+++ cronjobs-entrypoint.sh
@@ -3,8 +3,8 @@ set -e
 
 echo "Running cronjobs"
 
-echo "*/2 * * * * /var/local/pontaj/crontabs/run_snmp.sh && /var/local/pontaj/crontabs/check_mac_addresses.py" >> crontab.tmp
-echo "0 6 * * 1-5 /var/local/pontaj/crontabs/run_clear_macs.sh" >> crontab.tmp
+echo "*/2 * * * * /var/local/pontaj/crontabs/run_snmp.sh && python /var/local/pontaj/manage.py check_insert_mac_addresses" >> crontab.tmp
+echo "0 6 * * 1-5 python /var/local/pontaj/manage.py clear_mac_addresses" >> crontab.tmp
 
 crontab crontab.tmp
 rm crontab.tmp
